[deps/cppjieba] v5.1.2 -> v5.3.1

+++ deps/cppjieba/deps/limonp/test/unittest/TMd5.cpp
@@ -29,13 +29,13 @@ const char* const TEST_STR_MD5[] = {
 TEST(Md5Test, Test1) {
   ASSERT_EQ(sizeof(DICT_FILE)/sizeof(DICT_FILE[0]), sizeof(DICT_FILE_MD5)/sizeof(DICT_FILE_MD5[0]));
   string tmp;
-  for (uint i = 0; i < sizeof(DICT_FILE)/sizeof(DICT_FILE[0]); i++) {
+  for (size_t i = 0; i < sizeof(DICT_FILE)/sizeof(DICT_FILE[0]); i++) {
     md5File(DICT_FILE[i], tmp);
     ASSERT_EQ(tmp, string(DICT_FILE_MD5[i]));
   }
 
   ASSERT_EQ(sizeof(TEST_STR)/sizeof(TEST_STR[0]), sizeof(TEST_STR_MD5)/sizeof(TEST_STR_MD5[0]));
-  for (uint i = 0; i < sizeof(TEST_STR)/sizeof(TEST_STR[0]); i++) {
+  for (size_t i = 0; i < sizeof(TEST_STR)/sizeof(TEST_STR[0]); i++) {
     md5String(TEST_STR[i], tmp);
     ASSERT_EQ(tmp, string(TEST_STR_MD5[i]));
   }
